pyproject.toml
[project]
name = "gi-data-platform"
version = "0.1.0"
description = "GI Data Lakehouse — Pipeline Evolia → Bronze → Silver → Gold (OVHcloud)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "Mahamadou COULIBALY", email = "mahamadou.coulibaly@interaction-groupe.com" }]

dependencies = [
    # Connexion ERP Evolia (SQL Server on-prem)
    "pyodbc>=5.0.0",
    # Object Storage S3 (OVHcloud / compatible AWS S3)
    "boto3>=1.34.0",
    # Transformation Silver (DuckDB lit Parquet/JSON depuis S3)
    "duckdb>=1.1.0",
    # Data Warehouse Gold (PostgreSQL managé OVHcloud)
    "psycopg2-binary>=2.9.9",
    # Chargement .env.local / .env
    "python-dotenv>=1.0.0",
    # API OVHcloud (monitoring, gestion ressources)
    "ovh>=1.0.0",
]

[project.optional-dependencies]
# Qualité des données — Great Expectations
quality = [
    "great-expectations>=0.18.0",
]
# Transformation avancée — dbt Phase 2+
dbt = [
    "dbt-postgres>=1.8.0",
]
# Catalogue — DataHub Phase 4
catalogue = [
    "acryl-datahub>=0.12.0",
]
# Tests & CI
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-boto3",
]
# Géocodage BAN — enrichissement adresses
geo = [
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

# ─── Ruff ─────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["scripts"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501",  # line-length géré par formatter
    "B008",  # function call in defaults (dataclass fields)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "F811"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─── Pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=scripts",
    "--cov-report=term-missing",
    "--cov-fail-under=60",
]
markers = [
    "unit: Tests unitaires — zéro I/O externe (RunMode.OFFLINE)",
    "integration: Tests d'intégration — nécessite .env.local (RunMode.PROBE)",
    "slow: Tests > 10s (full-load, transformations volumineuses)",
]
filterwarnings = [
    "ignore::DeprecationWarning:boto3",
    "ignore::DeprecationWarning:duckdb",
]

# ─── Mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true   # boto3, pyodbc stubs manquants
exclude = ["dbt/", "tests/"]

# ─── Coverage ─────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["scripts"]
omit = [
    "scripts/probe_ddl*.py",
    "scripts/*__pycache__*",
]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
    "raise NotImplementedError",
    "\\.\\.\\.",
]

# ─── Build ────────────────────────────────────────────────────────────────────
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]